+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - **casePolice** config: autofix is now disabled by default.
 - **react** config: `configTypeAwareRules` now uses `files` and `ignores` from `ts` config by default.
 - **yaml** config: add a new config to target GitHub Actions workflow files, currently only disabling [`no-empty-mapping-value`](https://ota-meshi.github.io/eslint-plugin-yml/rules/no-empty-mapping-value.html) rule.
+- **vue** config: support Nuxt 4 new directory structure.
 - Added a new root option, `defaultConfigsStatus`, to control what configs are enabled or disabled by default.
 
 ## 0.10.0